configure: determine pointer-sized PNT_ARITH_TYPE
Compilers *really* dislike casts between pointers and integers not
matching them in size, so the simple approach of casting pointers to the
largest type available leads to warnings on 32-bit systems.  Try harder
to define PNT_ARITH_TYPE to a type with sizeof() == sizeof(void*)
instead of opting for a 64-bit type everywhere.  This should silence
the warnings on 32-bit platforms.