return: Fix PowerPC assembly
The original assembly used suspicious syntax. However, due to the
!defined(__OPTIMIZE__) guard this code was rarely built.

There nothing to stop the compiler using r0 between the two asm blocks,
which may have been the cause of the note mentioning it failed when
build with optimisation enabled.

Write a single asm statement that places the result in the given
location.

This builds for powerpc64le and passes tests.