Squash build races between dmalloc.h and .o files
This change modifies the object code to depend on dmalloc.h as it's
generated over the course of the build. This unbreaks parallel builds
(-j greater than 1).

NB: technically all of the local header dependencies should be fully
mapped to their .o files to ensure that the .o files are properly
rebuilt when the headers are modified, but this covers the core
concurrent build issue that was occurring prior to this change.

Closes: #119
Co-authored-by:	Bryan Drewery <bdrewery@FreeBSD.org>
Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>